typo

+++ BreathDetection/BreathDetection/ADCensus-kernel.cl.c
@@ -33,7 +33,7 @@ __kernel void kComputeCosts(__global uchar* L,
 }
 
 __kernel void kGetDisparityMap(__global float* costs,
-							   __global uchar* disp) {
+							   __global float* disp) {
 /*  find minimal costs values and construct isparity map */
 
 	const int2 xy = (get_global_id(0), get_global_id(1));
